pyproject.toml
[tool.poetry]
name = "ubx_rtk_base"
version = "1.0.0"
description = ""
authors = ["Oidlichtnwoada"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyserial = "^3.5"
pyubx2 = "^1.2.38"
pynmeagps = "^1.0.33"

[tool.poetry.scripts]
main = "ubx_rtk_base.main:main"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
mypy = "^1.8.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
